+++ pywr/nodes.py
@@ -498,6 +498,92 @@ class VirtualStorage(Drawable, _core.VirtualStorage, metaclass=NodeMeta):
         node = cls(model, nodes=nodes, **data)
         return node
 
+
+class RollingVirtualStorage(Drawable, _core.RollingVirtualStorage, metaclass=NodeMeta):
+    """A rolling virtual storage node useful for implementing rolling licences.
+
+    Parameters
+    ----------
+    model: pywr.core.Model
+    name: str
+        The name of the virtual node
+    nodes: list of nodes
+        List of inflow/out flow nodes that affect the storage volume
+    factors: list of floats
+        List of factors to multiply node flow by. Positive factors remove
+        water from the storage, negative factors remove it.
+    min_volume: float or parameter
+        The minimum volume the storage is allowed to reach.
+    max_volume: float or parameter
+        The maximum volume of the storage.
+    initial_volume: float
+        The initial storage volume.
+    timesteps : int
+        The number of timesteps to apply to the rolling storage over.
+    days : int
+        The number of days to apply the rolling storage over. Specifying a number of days (instead of a number
+        of timesteps) is only valid with models running a timestep of daily frequency.
+    cost: float or parameter
+        The cost of flow into/outfrom the storage.
+
+    Notes
+    -----
+    TODO: The cost property is not currently respected. See issue #242.
+    """
+    def __init__(self, model, name, nodes, **kwargs):
+        min_volume = pop_kwarg_parameter(kwargs, 'min_volume', 0.0)
+        if min_volume is None:
+            min_volume = 0.0
+        max_volume = pop_kwarg_parameter(kwargs, 'max_volume', 0.0)
+        initial_volume = kwargs.pop('initial_volume', 0.0)
+        cost = pop_kwarg_parameter(kwargs, 'cost', 0.0)
+        factors = kwargs.pop('factors', None)
+        days = kwargs.pop('days', None)
+        timesteps = kwargs.pop('timesteps', 0)
+
+        if not timesteps and not days:
+            raise ValueError("Either `timesteps` or `days` must be specified.")
+
+        super().__init__(model, name, **kwargs)
+
+        self.min_volume = min_volume
+        self.max_volume = max_volume
+        self.initial_volume = initial_volume
+        self.cost = cost
+        self.nodes = nodes
+        self.days = days
+        self.timesteps = timesteps
+
+        if factors is None:
+            self.factors = [1.0 for i in range(len(nodes))]
+        else:
+            self.factors = factors
+
+    def check(self):
+        super().check()
+        if self.cost not in (0.0, None):
+            raise NotImplementedError("RollingVirtualStorage does not currently support a non-zero cost.")
+
+    def setup(self, model):
+        if self.days is not None and self.days > 0:
+            try:
+                self.timesteps = self.days // self.model.timestepper.delta
+            except TypeError:
+                raise TypeError('A rolling period defined as a number of days is only valid with daily time-steps.')
+        if self.timesteps < 1:
+            raise ValueError('The number of time-steps for a RollingVirtualStorage node must be greater than one.')
+        super().setup(model)
+
+    @classmethod
+    def load(cls, data, model):
+        del(data["type"])
+        nodes = []
+        for node_name in data.pop("nodes"):
+            nodes.append(model._get_node_from_ref(model, node_name))
+        node = cls(model, nodes=nodes, **data)
+        return node
+
+
 class AnnualVirtualStorage(VirtualStorage):
     """A virtual storage which resets annually, useful for licences
 
@@ -537,6 +623,65 @@ class AnnualVirtualStorage(VirtualStorage):
                 # Reset to maximum volume (i.e. full capacity. )
                 self._reset_storage_only(use_initial_volume=self.reset_to_initial_volume)
                 self._last_reset_year = ts.year
+                self.active = True
+
+
+class SeasonalVirtualStorage(AnnualVirtualStorage):
+    """A virtual storage node that operates only for a specified period within a year.
+
+    This node is most useful for representing licences that are only enforced during specified periods. The
+    `reset_day` and `reset_month` parameters indicate when the node starts operating and the `end_day` and `end_month`
+    when it stops operating. For the period when the node is not operating, the volume of the node remains unchanged
+    and the node does not apply any constraints to the model.
+
+    The end_day and end_month can represent a date earlier in the year that the reset_day and and reset_month. This
+    situation represents a licence that operates across a year boundary. For example, one that is active between
+    October and March and not active between April and September.
+
+    Parameters
+    ----------
+    reset_day : int
+        The day of the month (0-31) when the node starts operating and its volume is reset to the initial value or
+        maximum volume.
+    reset_month : int
+        The month of the year (0-12) when the node starts operating and its volume is reset to the initial value or
+        maximum volume.
+    reset_to_initial_volume : bool
+        Reset the volume to the initial volume instead of maximum volume each year (default is False).
+    end_day : int
+        The day of the month (0-31) when the node stops operating.
+    end_month : int
+        The month of the year (0-12) when the node stops operating.
+    """
+
+    def __init__(self, *args, **kwargs):
+        self.end_day = kwargs.pop('end_day', 31)
+        self.end_month = kwargs.pop('end_month', 12)
+        self._last_active_year = None
+
+        super().__init__(*args, **kwargs)
+
+    def before(self, ts):
+        super().before(ts)
+
+        if ts.year != self._last_active_year:
+            if ts.index == 0:
+                if self._last_reset_year == ts.year:
+                    # First timestep is later in year than reset date
+                    if self.end_month < self.reset_month or \
+                            (self.end_month == self.reset_month and self.end_day <= self.reset_day):
+                        # end date is earlier in year than reset date so do not deactivate node in first year
+                        self._last_active_year = ts.year
+                else:
+                    # First timestep is earlier in year than reset date
+                    if self.end_month > self.reset_month or \
+                            (self.end_month == self.reset_month and self.end_day >= self.reset_day):
+                        # end date is later in year than reset date so node needs to be deactivated
+                        self.active = False
+            elif ts.month > self.end_month or \
+                    (ts.month == self.end_month and ts.day >= self.end_day):
+                self._last_active_year = ts.year
+                self.active = False
 
 
 class PiecewiseLink(Node):
@@ -717,7 +862,7 @@ class MultiSplitLink(PiecewiseLink):
             self._extra_inputs.append(inpt)
             self._extra_outputs.append(otpt)
 
-        # Now create an aggregated node for addition constaints if required.
+        # Now create an aggregated node for addition constraints if required.
         if factors is not None:
             if extra_slots+1 != len(factors):
                 raise ValueError("factors must have a length equal to extra_slots.")
@@ -743,8 +888,18 @@ class MultiSplitLink(PiecewiseLink):
         else:
             yield self.output
 
+    @classmethod
+    def load(cls, data, model):
+        # max_flow and cost should be lists of parameter definitions
+        max_flow = [load_parameter(model, p) for p in data.pop('max_flow')]
+        cost = [load_parameter(model, p) for p in data.pop('cost')]
+        factors = AggregatedNode.load_factors(model, data)
 
-class AggregatedStorage( Drawable, _core.AggregatedStorage, metaclass=NodeMeta):
+        del(data["type"])
+        return cls(model, max_flow=max_flow, cost=cost, factors=factors, **data)
+
+
+class AggregatedStorage(Drawable, _core.AggregatedStorage, metaclass=NodeMeta):
     """ An aggregated sum of other `Storage` nodes
 
     This object should behave like `Storage` by returning current `flow`, `volume` and `current_pc`.